backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.0
azure-identity==1.15.0
openai==1.12.0
pydantic==2.6.0
gunicorn==21.2.0
